implement file attachment, add File DTO [#504855]

+++ src/main/java/ru/job4j/service/SimpleCandidateService.java
@@ -2,7 +2,9 @@ package ru.job4j.service;
 
 import net.jcip.annotations.ThreadSafe;
 import org.springframework.stereotype.Service;
+import ru.job4j.dto.FileDto;
 import ru.job4j.model.Candidate;
+import ru.job4j.model.Vacancy;
 import ru.job4j.repository.CandidateRepository;
 
 import java.util.Collection;
@@ -12,24 +14,46 @@ import java.util.Optional;
 @ThreadSafe
 public class SimpleCandidateService implements CandidateService {
     private final CandidateRepository candidateRepository;
+    private final FileService fileService;
 
-    public SimpleCandidateService(CandidateRepository candidateRepository) {
+    public SimpleCandidateService(CandidateRepository candidateRepository, FileService fileService) {
         this.candidateRepository = candidateRepository;
+        this.fileService = fileService;
     }
 
     @Override
-    public Candidate save(Candidate candidate) {
+    public Candidate save(Candidate candidate, FileDto image) {
+        saveNewFile(candidate, image);
         return candidateRepository.save(candidate);
+    }
+
+    private void saveNewFile(Candidate candidate, FileDto image) {
+        var file = fileService.save(image);
+        candidate.setFileId(file.getId());
     }
 
     @Override
     public boolean deleteById(int id) {
-        return candidateRepository.deleteById(id);
+        var fileOptional = findById(id);
+        boolean isVacancyDeleted = false;
+        if (fileOptional.isPresent()) {
+            isVacancyDeleted = candidateRepository.deleteById(id);
+            fileService.deleteById(fileOptional.get().getFileId());
+        }
+        return isVacancyDeleted;
     }
 
     @Override
-    public boolean update(Candidate candidate) {
-        return candidateRepository.update(candidate);
+    public boolean update(Candidate candidate, FileDto image) {
+        var isNewFileEmpty = image.getContent().length == 0;
+        if (isNewFileEmpty) {
+            return candidateRepository.update(candidate);
+        }
+        var oldFileId = candidate.getFileId();
+        saveNewFile(candidate, image);
+        var isUpdated = candidateRepository.update(candidate);
+        fileService.deleteById(oldFileId);
+        return isUpdated;
     }
 
     @Override
