+++ src/main/java/ru/job4j/service/SimpleVacancyService.java
@@ -2,6 +2,7 @@ package ru.job4j.service;
 
 import net.jcip.annotations.ThreadSafe;
 import org.springframework.stereotype.Service;
+import ru.job4j.dto.FileDto;
 import ru.job4j.model.Vacancy;
 import ru.job4j.repository.VacancyRepository;
 
@@ -12,24 +13,46 @@ import java.util.Optional;
 @ThreadSafe
 public class SimpleVacancyService implements VacancyService {
     private final VacancyRepository vacancyRepository;
+    private final FileService fileService;
 
-    public SimpleVacancyService(VacancyRepository vacancyRepository) {
+    public SimpleVacancyService(VacancyRepository vacancyRepository, FileService fileService) {
         this.vacancyRepository = vacancyRepository;
+        this.fileService = fileService;
     }
 
     @Override
-    public Vacancy save(Vacancy vacancy) {
+    public Vacancy save(Vacancy vacancy, FileDto image) {
+        saveNewFile(vacancy, image);
         return vacancyRepository.save(vacancy);
+    }
+
+    private void saveNewFile(Vacancy vacancy, FileDto image) {
+        var file = fileService.save(image);
+        vacancy.setFileId(file.getId());
     }
 
     @Override
     public boolean deleteById(int id) {
-        return vacancyRepository.deleteById(id);
+        var fileOptional = findById(id);
+        boolean isVacancyDeleted = false;
+        if (fileOptional.isPresent()) {
+            isVacancyDeleted = vacancyRepository.deleteById(id);
+            fileService.deleteById(fileOptional.get().getFileId());
+        }
+        return isVacancyDeleted;
     }
 
     @Override
-    public boolean update(Vacancy vacancy) {
-        return vacancyRepository.update(vacancy);
+    public boolean update(Vacancy vacancy, FileDto image) {
+        var isNewFileEmpty = image.getContent().length == 0;
+        if (isNewFileEmpty) {
+            return vacancyRepository.update(vacancy);
+        }
+        var oldFileId = vacancy.getFileId();
+        saveNewFile(vacancy, image);
+        var isUpdated = vacancyRepository.update(vacancy);
+        fileService.deleteById(oldFileId);
+        return isUpdated;
     }
 
     @Override
